Moving core functions from pmodel.functions
